Complete the initial separation
