+++ cpp/src/cufile/driver.cpp
@@ -23,7 +23,7 @@
 #include <kvikio/shim/cufile_h_wrapper.hpp>
 
 namespace kvikio {
-namespace detail {
+namespace {
 
 [[nodiscard]] inline bool get_driver_flag(unsigned int prop, unsigned int flag) noexcept
 {
@@ -38,7 +38,7 @@ inline void set_driver_flag(unsigned int& prop, unsigned int flag, bool val) noe
     prop &= ~(1U << flag);
   }
 }
-}  // namespace detail
+}  // namespace
 
 #ifdef KVIKIO_CUFILE_FOUND
 
@@ -70,31 +70,31 @@ bool DriverProperties::is_gds_available()
   return !(get_nvfs_major_version() == 0 && get_nvfs_minor_version() == 0);
 }
 
-[[nodiscard]] unsigned int DriverProperties::get_nvfs_major_version()
+unsigned int DriverProperties::get_nvfs_major_version()
 {
   lazy_init();
   return _props.nvfs.major_version;
 }
 
-[[nodiscard]] unsigned int DriverProperties::get_nvfs_minor_version()
+unsigned int DriverProperties::get_nvfs_minor_version()
 {
   lazy_init();
   return _props.nvfs.minor_version;
 }
 
-[[nodiscard]] bool DriverProperties::get_nvfs_allow_compat_mode()
+bool DriverProperties::get_nvfs_allow_compat_mode()
 {
   lazy_init();
-  return detail::get_driver_flag(_props.nvfs.dcontrolflags, CU_FILE_ALLOW_COMPAT_MODE);
+  return get_driver_flag(_props.nvfs.dcontrolflags, CU_FILE_ALLOW_COMPAT_MODE);
 }
 
-[[nodiscard]] bool DriverProperties::get_nvfs_poll_mode()
+bool DriverProperties::get_nvfs_poll_mode()
 {
   lazy_init();
-  return detail::get_driver_flag(_props.nvfs.dcontrolflags, CU_FILE_USE_POLL_MODE);
+  return get_driver_flag(_props.nvfs.dcontrolflags, CU_FILE_USE_POLL_MODE);
 }
 
-[[nodiscard]] std::size_t DriverProperties::get_nvfs_poll_thresh_size()
+std::size_t DriverProperties::get_nvfs_poll_thresh_size()
 {
   lazy_init();
   return _props.nvfs.poll_thresh_size;
@@ -104,7 +104,7 @@ void DriverProperties::set_nvfs_poll_mode(bool enable)
 {
   lazy_init();
   CUFILE_TRY(cuFileAPI::instance().DriverSetPollMode(enable, get_nvfs_poll_thresh_size()));
-  detail::set_driver_flag(_props.nvfs.dcontrolflags, CU_FILE_USE_POLL_MODE, enable);
+  set_driver_flag(_props.nvfs.dcontrolflags, CU_FILE_USE_POLL_MODE, enable);
 }
 
 void DriverProperties::set_nvfs_poll_thresh_size(std::size_t size_in_kb)
@@ -114,20 +114,20 @@ void DriverProperties::set_nvfs_poll_thresh_size(std::size_t size_in_kb)
   _props.nvfs.poll_thresh_size = size_in_kb;
 }
 
-[[nodiscard]] std::vector<CUfileDriverControlFlags> DriverProperties::get_nvfs_statusflags()
+std::vector<CUfileDriverControlFlags> DriverProperties::get_nvfs_statusflags()
 {
   lazy_init();
   std::vector<CUfileDriverControlFlags> ret;
-  if (detail::get_driver_flag(_props.nvfs.dcontrolflags, CU_FILE_USE_POLL_MODE)) {
+  if (get_driver_flag(_props.nvfs.dcontrolflags, CU_FILE_USE_POLL_MODE)) {
     ret.push_back(CU_FILE_USE_POLL_MODE);
   }
-  if (detail::get_driver_flag(_props.nvfs.dcontrolflags, CU_FILE_ALLOW_COMPAT_MODE)) {
+  if (get_driver_flag(_props.nvfs.dcontrolflags, CU_FILE_ALLOW_COMPAT_MODE)) {
     ret.push_back(CU_FILE_ALLOW_COMPAT_MODE);
   }
   return ret;
 }
 
-[[nodiscard]] std::size_t DriverProperties::get_max_device_cache_size()
+std::size_t DriverProperties::get_max_device_cache_size()
 {
   lazy_init();
   return _props.max_device_cache_size;
@@ -140,13 +140,13 @@ void DriverProperties::set_max_device_cache_size(std::size_t size_in_kb)
   _props.max_device_cache_size = size_in_kb;
 }
 
-[[nodiscard]] std::size_t DriverProperties::get_per_buffer_cache_size()
+std::size_t DriverProperties::get_per_buffer_cache_size()
 {
   lazy_init();
   return _props.per_buffer_cache_size;
 }
 
-[[nodiscard]] std::size_t DriverProperties::get_max_pinned_memory_size()
+std::size_t DriverProperties::get_max_pinned_memory_size()
 {
   lazy_init();
   return _props.max_device_pinned_mem_size;
@@ -159,7 +159,7 @@ void DriverProperties::set_max_pinned_memory_size(std::size_t size_in_kb)
   _props.max_device_pinned_mem_size = size_in_kb;
 }
 
-[[nodiscard]] std::size_t DriverProperties::get_max_batch_io_size()
+std::size_t DriverProperties::get_max_batch_io_size()
 {
 #ifdef KVIKIO_CUFILE_BATCH_API_FOUND
   lazy_init();
@@ -176,27 +176,27 @@ DriverProperties::DriverProperties() {}
 
 bool DriverProperties::is_gds_available() { return false; }
 
-[[nodiscard]] unsigned int DriverProperties::get_nvfs_major_version()
+unsigned int DriverProperties::get_nvfs_major_version()
 {
   throw CUfileException("KvikIO not compiled with cuFile.h");
 }
 
-[[nodiscard]] unsigned int DriverProperties::get_nvfs_minor_version()
+unsigned int DriverProperties::get_nvfs_minor_version()
 {
   throw CUfileException("KvikIO not compiled with cuFile.h");
 }
 
-[[nodiscard]] bool DriverProperties::get_nvfs_allow_compat_mode()
+bool DriverProperties::get_nvfs_allow_compat_mode()
 {
   throw CUfileException("KvikIO not compiled with cuFile.h");
 }
 
-[[nodiscard]] bool DriverProperties::get_nvfs_poll_mode()
+bool DriverProperties::get_nvfs_poll_mode()
 {
   throw CUfileException("KvikIO not compiled with cuFile.h");
 }
 
-[[nodiscard]] std::size_t DriverProperties::get_nvfs_poll_thresh_size()
+std::size_t DriverProperties::get_nvfs_poll_thresh_size()
 {
   throw CUfileException("KvikIO not compiled with cuFile.h");
 }
@@ -211,12 +211,12 @@ void DriverProperties::set_nvfs_poll_thresh_size(std::size_t size_in_kb)
   throw CUfileException("KvikIO not compiled with cuFile.h");
 }
 
-[[nodiscard]] std::vector<CUfileDriverControlFlags> DriverProperties::get_nvfs_statusflags()
+std::vector<CUfileDriverControlFlags> DriverProperties::get_nvfs_statusflags()
 {
   throw CUfileException("KvikIO not compiled with cuFile.h");
 }
 
-[[nodiscard]] std::size_t DriverProperties::get_max_device_cache_size()
+std::size_t DriverProperties::get_max_device_cache_size()
 {
   throw CUfileException("KvikIO not compiled with cuFile.h");
 }
@@ -226,12 +226,12 @@ void DriverProperties::set_max_device_cache_size(std::size_t size_in_kb)
   throw CUfileException("KvikIO not compiled with cuFile.h");
 }
 
-[[nodiscard]] std::size_t DriverProperties::get_per_buffer_cache_size()
+std::size_t DriverProperties::get_per_buffer_cache_size()
 {
   throw CUfileException("KvikIO not compiled with cuFile.h");
 }
 
-[[nodiscard]] std::size_t DriverProperties::get_max_pinned_memory_size()
+std::size_t DriverProperties::get_max_pinned_memory_size()
 {
   throw CUfileException("KvikIO not compiled with cuFile.h");
 }
@@ -241,7 +241,7 @@ void DriverProperties::set_max_pinned_memory_size(std::size_t size_in_kb)
   throw CUfileException("KvikIO not compiled with cuFile.h");
 }
 
-[[nodiscard]] std::size_t DriverProperties::get_max_batch_io_size()
+std::size_t DriverProperties::get_max_batch_io_size()
 {
   throw CUfileException("KvikIO not compiled with cuFile.h");
 }
